requirements.txt
numpy
pandas
seaborn
matplotlib
scikit-learn
catboost
xgboost
dill
python-dotenv
flask-cors
googlesearch-python
tld
-e .